Cluster starts with inlinemanifest cilium
