+++ tasks_pkg/__init__.py
@@ -2,3 +2,4 @@ from . import apps
 from . import helpers
 from . import metal
 from . import network
+from . import cluster
